requirements.txt
greenlet==3.0.3
grpcio==1.66.1
grpcio-tools==1.66.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
protobuf==5.28.0
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-dotenv==0.5.2
python-dotenv==1.0.1
setuptools==74.1.1
SQLAlchemy==2.0.33
typing_extensions==4.12.2
